+++ test/unit/gen_class/unit_test_run3.sh
@@ -5,7 +5,6 @@
 #   is located at.
 
 echo "Unit test: gen_class module"  
-/usr/bin/python3 ./test/unit/gen_class/get_inst.py
 /usr/bin/python3 ./test/unit/gen_class/setup_mail.py
 /usr/bin/python3 ./test/unit/gen_class/argparser_arg_add_def.py
 /usr/bin/python3 ./test/unit/gen_class/argparser_arg_cond_req.py