+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [3.0.4] - 2024-08-20
+- Removed the get_inst function.
+
+### Changed
+- gen_class.Daemon.stop, gen_class.Daemon2.stop, gen_class.Mail.read_stdin, gen_Class.Mail.send_mail, gen_Class.Mail.send_mailx, gen_libs.make_md5_hash, gen_libs.compress: Removed get_inst call.
+
+### Removed
+- gen_class.get_inst
+- gen_libs.get_inst
 
 
 ## [3.0.3] - 2024-02-21